Coordinate up and running - replacing Vector for typically publicly exposed properties such as Position on Nodes and Target on Camera.
The Coordinate type has a bunch of implicit conversions between Coordinate and Vector, so one can actually work with them close to as if they were the same type.

The reason for introducing this, is basically Xaml support. Version 0.8.8.0 introduced a whole set of controls specifically designed to tackle the Xaml problem. All though they work, they represent a bunch of classes that needs maintaining anytime base classes change, as they are only referencing the ones in Balder.Core and not actually inheriting them. Instead, the Property<> class used in Core now allows for being platform ignorant in the property definition and then implemented specific support within that type for things like DependencyProperty.