Fixed coloring for child nodes after doing some changes in how Mesh works. It doesn't render itself anymore, since all its doing is loading geometries and geometries already knnow how to render themselves.
need to revisit this though, since color is also something that is coming from the files.

But basic rule will be in the future: Colors set in the file will at load time be applied, if one specifies color in the Xaml or programatically this will override the color on the node itself but also on any children.